removed cache_manager to separate the redis implementation from the main functionality
